+++ frontend/src/pages/GeneralSettings/EmbeddingPreference/index.jsx
@@ -19,7 +19,7 @@ import EmbedderItem from "@/components/EmbeddingSelection/EmbedderItem";
 import { CaretUpDown, MagnifyingGlass, X } from "@phosphor-icons/react";
 import { useModal } from "@/hooks/useModal";
 import ModalWrapper from "@/components/ModalWrapper";
-import CTAButton from "@/components/CTAButton";
+import CTAButton from "@/components/lib/CTAButton";
 
 export default function GeneralEmbeddingPreference() {
   const [saving, setSaving] = useState(false);
