+++ frontend/src/pages/GeneralSettings/LLMPreference/index.jsx
@@ -35,7 +35,7 @@ import GroqAiOptions from "@/components/LLMSelection/GroqAiOptions";
 
 import LLMItem from "@/components/LLMSelection/LLMItem";
 import { CaretUpDown, MagnifyingGlass, X } from "@phosphor-icons/react";
-import CTAButton from "@/components/CTAButton";
+import CTAButton from "@/components/lib/CTAButton";
 
 export default function GeneralLLMPreference() {
   const [saving, setSaving] = useState(false);
