+++ frontend/src/pages/WorkspaceSettings/Members/index.jsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from "react";
 import * as Skeleton from "react-loading-skeleton";
 import AddMemberModal from "./AddMemberModal";
 import WorkspaceMemberRow from "./WorkspaceMemberRow";
-import CTAButton from "@/components/CTAButton";
+import CTAButton from "@/components/lib/CTAButton";
 
 export default function Members({ workspace }) {
   const [loading, setLoading] = useState(true);
